requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
numpy==1.26.4
pillow==10.4.0
tensorflow==2.13.1
llama-cpp-python==0.2.78
gunicorn==21.2.0


